requirements.txt
wordcloud
pillow
numpy
